MLProject/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
mlflow==2.19.0
python-dotenv
dagshub
